feat: finish the acceptance test
We need to mock the date in order to reproduce when the customer inserts
a deposit in different days. To verify that the service work well and
the output is correct, we need to create a output service to pass it to
the accountService. The output service will be in charge of printing the
data.